\name{mPathy-package}
\alias{mPathy-package}
\alias{mPathy}
\docType{package}
\title{
\packageTitle{mPathy}
}
\description{
\packageDescription{mPathy}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{mPathy}
\packageIndices{mPathy}
mPathy provides means of fast pathway reconstruction that is it provides visual by means of KEGG pathway maps. The maps are interactive and KEGG orthologs are highlighted according to selected taxonomy IDs. Phylo-functional annotation of reads is provided by uProc.
}
\author{
\packageAuthor{mPathy}

Maintainer: \packageMaintainer{mPathy}
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC102409/}
}
\keyword{metagenomics}
\seealso{
\code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{
TODO
}
